+++ component/grpc/component_test.go
@@ -77,8 +77,7 @@ func TestComponent_Run_Unary(t *testing.T) {
 		assert.NoError(t, cmp.Run(ctx))
 		chDone <- struct{}{}
 	}()
-	conn, err := grpc.DialContext(ctx, "localhost:60000", grpc.WithTransportCredentials(insecure.NewCredentials()),
-		grpc.WithBlock())
+	conn, err := grpc.NewClient("localhost:60000", grpc.WithTransportCredentials(insecure.NewCredentials()))
 	require.NoError(t, err)
 	c := examples.NewGreeterClient(conn)
 
@@ -162,8 +161,8 @@ func TestComponent_Run_Stream(t *testing.T) {
 		assert.NoError(t, cmp.Run(ctx))
 		chDone <- struct{}{}
 	}()
-	conn, err := grpc.DialContext(ctx, "localhost:60000", grpc.WithTransportCredentials(insecure.NewCredentials()),
-		grpc.WithBlock())
+
+	conn, err := grpc.NewClient("localhost:60000", grpc.WithTransportCredentials(insecure.NewCredentials()))
 	require.NoError(t, err)
 	c := examples.NewGreeterClient(conn)
