+++ examples/client/main.go
@@ -131,7 +131,7 @@ func sendHTTPRequest(ctx context.Context) error {
 }
 
 func sendGRPCRequest(ctx context.Context) error {
-	cc, err := patrongrpc.DialContext(ctx, examples.GRPCTarget, grpc.WithTransportCredentials(insecure.NewCredentials()))
+	cc, err := patrongrpc.NewClient(examples.GRPCTarget, grpc.WithTransportCredentials(insecure.NewCredentials()))
 	if err != nil {
 		return err
 	}
